moved panorama and prisma underneath inventory
