+++ frontend/src/app/inventory/panorama/panorama.component.html
