+++ frontend/src/app/inventory/prisma/prisma.component.html
