+++ frontend/src/app/inventory/prisma/prisma.component.scss
@@ -1,0 +1,3 @@
+.spacer {
+  flex: 1 1 auto;
+}
