+++ frontend/src/app/inventory/prisma/prisma.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { AuthService } from '../auth.service';
+import { AuthService } from '../../auth.service';
 import { catchError, map } from 'rxjs/operators';
 import { of } from 'rxjs';
