--- frontend/src/app/prisma/prisma.component.scss
@@ -1,4 +1,0 @@
-.spacer {
-    flex: 1 1 auto;
-  }
-
